Improve last info class
