+++ spec/Phpro/SoapClient/Soap/HttpBinding/LastRequestInfoSpec.php
@@ -2,9 +2,6 @@
 
 namespace spec\Phpro\SoapClient\Soap\HttpBinding;
 
-use GuzzleHttp\Psr7\Request;
-use GuzzleHttp\Psr7\Response;
-use Phpro\SoapClient\Soap\SoapClient;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Phpro\SoapClient\Soap\HttpBinding\LastRequestInfo;
@@ -51,41 +48,6 @@ class LastRequestInfoSpec extends ObjectBehavior
         $result->getLastRequestHeaders()->shouldBe('');
         $result->getLastRequest()->shouldBe('');
         $result->getLastResponseHeaders()->shouldBe('');
-        $result->getLastResponse()->shouldBe('');
-    }
-
-    // Note: the __get* cannot be mocked with phpspec.
-    function it_can_create_from_a_soapclient(\SoapClient $client)
-    {
-        $result = $this->createFromSoapClient($client);
-        $result->shouldBeAnInstanceOf(LastRequestInfo::class);
-        $result->getLastRequestHeaders()->shouldBe('');
-        $result->getLastRequest()->shouldBe('');
-        $result->getLastResponseHeaders()->shouldBe('');
-        $result->getLastResponse()->shouldBe('');
-    }
-
-    function it_can_load_from_psr7_request_and_response()
-    {
-        $request = new Request('POST', '/', ['x-request-header' => 'value'], 'REQUESTBODY');
-        $response = new Response(200, ['x-response-header' => 'value'], 'RESPONSEBODY');
-
-        $result = $this->createFromPsr7RequestAndResponse($request, $response);
-        $result->getLastRequestHeaders()->shouldBe('x-request-header: value');
-        $result->getLastRequest()->shouldBe('REQUESTBODY');
-        $result->getLastResponseHeaders()->shouldBe('x-response-header: value');
-        $result->getLastResponse()->shouldBe('RESPONSEBODY');
-    }
-
-    function it_can_load_from_psr7_request_and_response_without_body()
-    {
-        $request = new Request('GET', '/', ['x-request-header' => 'value'], '');
-        $respone = new Response(204, ['x-response-header' => 'value'], '');
-
-        $result = $this->createFromPsr7RequestAndResponse($request, $respone);
-        $result->getLastRequestHeaders()->shouldBe('x-request-header: value');
-        $result->getLastRequest()->shouldBe('');
-        $result->getLastResponseHeaders()->shouldBe('x-response-header: value');
         $result->getLastResponse()->shouldBe('');
     }
 }
